refactor: organize driver classes into subpackages
