+++ src/main/java/com/example/aqa/configuration/MainConfiguration.java
@@ -8,11 +8,11 @@ import com.example.aqa.configuration.driver.waiter.WebDriverWaitConfiguration;
 import com.example.aqa.configuration.extension.ExtensionConfiguration;
 import com.example.aqa.configuration.rest.RestApiClientConfiguration;
 import com.example.aqa.configuration.driver.appium.AppiumConfiguration;
-import com.example.aqa.driver.AppDriver;
-import com.example.aqa.driver.AppiumBasedAppDriver;
-import com.example.aqa.driver.MockAppDriver;
-import com.example.aqa.driver.SeleniumBasedAppDriver;
-import com.example.aqa.driver.PlaywrightBasedAppDriver;
+import com.example.aqa.driver.core.AppDriver;
+import com.example.aqa.driver.appium.AppiumBasedAppDriver;
+import com.example.aqa.driver.mock.MockAppDriver;
+import com.example.aqa.driver.selenium.SeleniumBasedAppDriver;
+import com.example.aqa.driver.playwright.PlaywrightBasedAppDriver;
 import com.microsoft.playwright.Page;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.support.ui.WebDriverWait;
